Old version of the application
I tried to use Room persistance library for the database but it turned out that I don't know how to use it, and there aren't any informative tutorials for that matter. I decided to abandon that idea and switch to good ol' SQLiteDatabaseHelper, but here's the old version just in case.